eco32: fix pushing of varargs to stack
When pushing the varargs that are held in register to the stack,
only the registers related to the varargs should be pushed,
not all registers.

This also fixes the calculation of the offsets at which the
registers should be pushed upon, they were wrong for
everything except the special case f (n, ...),
i.e. f (a, b, n, ...) would produce erronous stack locations.

A couple of whitespace and coding style fixes are snuck
into this as well.