eco32: split args between regs and stack
This is against the official eco32 abi that doesn't allow
splitting of args between regs and stack.
But when the args are not split it creates a hard to solve
case when an argument that doesn't fit into the last argument
register is passed to a variadic function.
In that case a "hole" is created in the stack where
the register is supposed to be, but the generic variadic
argument handler expects the arguments to be continous on
the stack.